we have motor driver working with PWM 16bit, and euler angles working. pid values are next to be found

+++ workspace_kk/DDL_FINAL_invert_pendulum/src/Pid.c
@@ -11,7 +11,7 @@ extern uint32_t millis();
 
 /*working variables*/
 unsigned long lastTime;
-double Input, Output, Setpoint;
+float Input, Output, Setpoint;
 double errSum, lastErr;
 double kp, ki, kd;
 void compute()
